move contracts to correct folder
